requirements.txt
django
djangorestframework
psycops2
django-cors-headers
coreapi